requirements.txt
Flask
Flask-SQLAlchemy
flask-marshmallow
eventlet
Flask-SocketIO
pyjwt==1.4.2
marshmallow-sqlalchemy
postgres
gunicorn
flask-bcrypt